+++ src/app/database/UserRepository.java
@@ -6,6 +6,9 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 
+/**
+ * Class to execute all CRUD queries for the Users to the database.
+ */
 public class UserRepository implements Repository<String, User> {
 
     private final ConnectionManager connection;
@@ -14,6 +17,13 @@ public class UserRepository implements Repository<String, User> {
         this.connection = connection;
     }
 
+    /**
+     * Execute the query for saving an User instance.
+     *
+     * @param user     An instance of the User class to be saved.
+     * @param userType A String that indicates which user types is
+     * @throws SQLException When an username was already registered.
+     */
     @Override
     public void create(User user, String userType) throws SQLException {
         String query = "INSERT INTO users (name, type, password) VALUES (?, ?, ?)";
@@ -24,6 +34,12 @@ public class UserRepository implements Repository<String, User> {
         statement.execute();
     }
 
+    /**
+     * Execute the query for reading all users from the database.
+     *
+     * @return A resultSet containing all user's fields.
+     * @throws SQLException When no users found.
+     */
     @Override
     public ResultSet readAll() throws SQLException {
         String query = "SELECT * FROM users";
@@ -31,6 +47,13 @@ public class UserRepository implements Repository<String, User> {
         return statement.executeQuery(query);
     }
 
+    /**
+     * Obtain an user from the database from his username.
+     *
+     * @param username A string that represents his username.
+     * @return A result set containing all his fields.
+     * @throws SQLException When any username match.
+     */
     @Override
     public ResultSet read(String username) throws SQLException {
         String query = "SELECT * FROM users WHERE name = ?";
@@ -39,6 +62,13 @@ public class UserRepository implements Repository<String, User> {
         return statement.executeQuery();
     }
 
+    /**
+     * Update an user info from the database depending of his username.
+     *
+     * @param username A String to find the user by his name.
+     * @param newData  An user object generated with his new information.
+     * @throws SQLException For a syntax error on SQL.
+     */
     @Override
     public void update(String username, User newData) throws SQLException {
         String query = "UPDATE users SET name = ?, password = ? WHERE name = ?";
@@ -49,6 +79,12 @@ public class UserRepository implements Repository<String, User> {
         statement.executeUpdate();
     }
 
+    /**
+     * Deleting a user row from the database by looking for his username.
+     *
+     * @param username A String to find the user by his name.
+     * @throws SQLException For a syntax error on SQL.
+     */
     @Override
     public void delete(String username) throws SQLException {
         String query = "DELETE FROM users WHERE name = ?";
