+++ src/app/models/proceedings/ProceedingsMetadata.java
@@ -1,0 +1,17 @@
+package app.models.proceedings;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProceedingsMetadata {
+
+    private String name;
+    private int docOrder;
+    private ProceedingsDate dates;
+    private PageData pageData;
+    private ExtraData extraData;
+}
