+++ src/app/views/Window.java
@@ -3,7 +3,6 @@ package app.views;
 import com.formdev.flatlaf.FlatLightLaf;
 import java.awt.*;
 import javax.swing.*;
-
 import lombok.Data;
 import lombok.EqualsAndHashCode;
